safari-600.7-branch/LayoutTests/crypto/subtle/aes-postMessage-expected.txt

Test sending crypto keys via postMessage.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.algorithm.length is 128
PASS key.usages is ["decrypt", "encrypt", "sign", "verify"]
PASS successfullyParsed is true

TEST COMPLETE
